updated statistic models
